requirements.txt
matplotlib==3.5.2
nibabel==3.2.2
numpy==1.22.4
progressbar==2.5
protobuf==3.19.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scipy==1.8.1
tensorboard==2.9.0
torch==1.12.1
tqdm==4.64.0
